[#88] AdminController 삭제
