+++ src/main/java/com/snack/news/controller/NewsController.java
@@ -5,11 +5,13 @@ import com.snack.news.dto.NewsDto;
 import com.snack.news.dto.WrappedResponse;
 import com.snack.news.service.NewsService;
 import lombok.AllArgsConstructor;
+import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.List;
+import java.util.Optional;
 
 @AllArgsConstructor
 @RestController
@@ -26,8 +28,23 @@ public class NewsController {
 		return WrappedResponse.ok(result);
 	}
 
+	@PostMapping
+	public ResponseEntity<NewsDto> createNews(@Valid @RequestBody NewsDto newsDto) {
+		return WrappedResponse.ok(newsService.createNews(newsDto));
+	}
+
 	@GetMapping("/{newsId}")
 	public ResponseEntity<News> getNews(@PathVariable Long newsId) {
 		return WrappedResponse.ok(newsService.getNews(newsId));
+	}
+
+	@GetMapping({"/news", "/news/{page}"})
+	public ResponseEntity<Page<News>> getNewsList(@SuppressWarnings("OptionalUsedAsFieldOrParameterType")
+												  @PathVariable Optional<Integer> page) {
+		Page<News> result = newsService.getNewsListForAdmin(page.orElse(0));
+		if (result.isEmpty()) {
+			return ResponseEntity.noContent().build();
+		}
+		return WrappedResponse.ok(result);
 	}
 }
